+++ src/main/java/subway/line/dao/LineMapper.java
@@ -1,19 +1,21 @@
-package subway.section;
+package subway.line.dao;
 
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
+import subway.line.domain.Line;
 
 import java.io.Serializable;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 @Component
-public class SectionMapper implements RowMapper<Section>, Serializable {
+public class LineMapper implements RowMapper<Line>, Serializable {
 
     @Override
-    public Section mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Section(rs.getLong("id"),
-                rs.getLong("line_id"),
+    public Line mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return new Line(rs.getLong("id"),
+                rs.getString("name"),
+                rs.getString("color"),
                 rs.getLong("up_station_id"),
                 rs.getLong("down_station_id"),
                 rs.getInt("distance"));
