+++ src/main/java/subway/section/domain/SectionRequest.java
@@ -1,4 +1,4 @@
-package subway.section;
+package subway.section.domain;
 
 public class SectionRequest {
     private Long upStationId;
