+++ src/main/java/subway/station/dao/StationDao.java
@@ -1,7 +1,8 @@
-package subway.station;
+package subway.station.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
+import subway.station.domain.Station;
 
 import javax.annotation.Resource;
 import java.util.List;
