+++ src/main/java/subway/station/domain/Station.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.domain;
 
 import java.util.Objects;
