+++ src/main/java/subway/station/domain/StationResponse.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.domain;
 
 public class StationResponse {
     private Long id;
